refactor: refactoring booking table and booking item to reuse in cart and account pages
