+++ src/app/account/pages/account-page/account-page.component.ts
@@ -1,7 +1,10 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
 import { logoutUser } from 'src/app/redux/actions/user.action';
+import { selectAllBookings } from 'src/app/redux/selectors/user.selectors';
+import { UserBooking } from 'src/app/shared/models/user.model';
 import { Paths } from 'src/app/types/enums';
 
 @Component({
@@ -11,6 +14,10 @@ import { Paths } from 'src/app/types/enums';
 })
 export class AccountPageComponent {
   public constructor(private store: Store, private router: Router) {}
+
+  public get allUserBookings$(): Observable<UserBooking[]> {
+    return this.store.select(selectAllBookings);
+  }
 
   public logout(): void {
     localStorage.removeItem('token');
