+++ src/app/shared/components/booking-table/booking-item/booking-item.component.scss
@@ -28,7 +28,7 @@ p {
 }
 
 .flight-number-container {
-  width: 25rem;
+  width: 28rem;
 }
 
 .way {
