+++ src/app/shared/components/booking-table/booking-table.component.ts
@@ -1,14 +1,10 @@
 /* eslint-disable no-return-assign */
-import { Component } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { Paths } from 'src/app/types/enums';
 import { Router } from '@angular/router';
 import { PaymentService } from 'src/app/core/services/payment.service';
-import { Store } from '@ngrx/store';
-
 import { Observable } from 'rxjs';
-import { selectUnpaidBookings } from 'src/app/redux/selectors/user.selectors';
 import { SelectedBookingService } from 'src/app/core/services/selected-booking.service';
-import { editBooking } from 'src/app/redux/actions/user.action';
 import { UserBooking } from '../../models/user.model';
 
 interface ChoosenBookings {
@@ -24,21 +20,16 @@ interface ChoosenBookings {
 export class BookingTableComponent {
   public selectedBookings: Array<ChoosenBookings> = [];
 
-  public isPopupVisible = false;
+  @Input() public userBookings$!: Observable<UserBooking[]>;
 
   public constructor(
     private router: Router,
     private paymentService: PaymentService,
-    private store: Store,
     private selectedBookingService: SelectedBookingService
   ) {}
 
   public get isAllChecked(): boolean {
     return this.selectedBookingService.getCurrentAllSelectedValue();
-  }
-
-  public get unpaidUserBookings$(): Observable<UserBooking[]> {
-    return this.store.select(selectUnpaidBookings);
   }
 
   public navToMain(): void {
@@ -54,10 +45,6 @@ export class BookingTableComponent {
     return sum;
   }
 
-  public get selectedCount(): number {
-    return this.selectedBookingService.bookings.length;
-  }
-
   public isAllBookingsSelected(bookings: UserBooking[]): void {
     this.selectedBookingService.changeAllSelectedValue();
     const isAllChecked =
@@ -69,26 +56,5 @@ export class BookingTableComponent {
     } else {
       this.selectedBookingService.clearBookings();
     }
-  }
-
-  public payBookings(): void {
-    const array = this.selectedBookingService.bookings.slice();
-    array.forEach(item => {
-      const booking: UserBooking = {
-        id: item.booking.id,
-        paid: true,
-        bookingInfo: item.booking.bookingInfo,
-        flights: item.booking.flights,
-        passengers: item.booking.passengers,
-      };
-      this.store.dispatch(editBooking({ bookings: booking }));
-    });
-    this.selectedBookingService.clearBookings();
-    this.selectedBookingService.changeAllSelectedValue();
-    this.togglePopup();
-  }
-
-  public togglePopup(): void {
-    this.isPopupVisible = !this.isPopupVisible;
   }
 }
