+++ src/app/shopping-cart/pages/cart-page.component.scss
@@ -12,3 +12,79 @@ p {
   font-weight: 700;
   line-height: 2.8rem;
 }
+
+.overlay {
+  position: fixed;
+  z-index: 5;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  background-color: var(--auxiliary-color-opacity02);
+}
+
+.apply-container {
+  padding: 4rem 3.5rem 0;
+  display: flex;
+  justify-content: space-between;
+}
+
+.promo-container {
+  width: 31rem;
+  padding: 0.4rem 1.6rem;
+  display: flex;
+  justify-content: space-between;
+  border: 1px solid var(--auxiliary-color-opacity02);
+  border-radius: 0.5rem;
+}
+
+.promo-apply {
+  font-size: 1.6rem;
+  font-weight: 500;
+  color: var(--primary-color);
+}
+
+.without-padding {
+  padding: 0;
+}
+
+.selected-container {
+  display: flex;
+  align-items: center;
+  gap: 3.8rem;
+}
+
+.select-content {
+  display: flex;
+  gap: 1rem;
+  color: var(--auxiliary-color);
+}
+
+.popup {
+  width: 32rem;
+  padding: 2rem;
+  display: none;
+  flex-direction: column;
+  gap: 5rem;
+  align-items: flex-end;
+  font-size: 1.8rem;
+  position: absolute;
+  top: 10vh;
+  left: calc(50% - 16rem);
+}
+
+.approve-text {
+  text-align: center;
+  font-weight: 400;
+  font-size: 2.2rem;
+  line-height: 3rem;
+}
+
+.pisioned-right {
+  width: 11rem;
+}
+
+.visible {
+  display: flex;
+  z-index: 20;
+}
